pyproject.toml
[project]
name = "envresolve"
description = "Resolve env vars from secret stores."
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "osoken"}]
dependencies = [
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
azure = [
    "azure-identity>=1.19.0",
    "azure-keyvault-secrets>=4.9.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "nox>=2025.5.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.13.3",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.21",
    "mkdocstrings>=0.30.1",
    "mkdocstrings-python>=1.18.2",
]
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "envresolve.__version__"}


[tool.ruff]
indent-width = 4
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",  # trailing-comma-missing (conflicts with formatter)
    "ISC001",  # single-line-implicit-string-concatenation (conflicts with formatter)
    "D203",    # one-blank-line-before-class (conflicts with D211)
    "D213",    # multi-line-summary-second-line (conflicts with D212)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert-used (pytest uses asserts)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests", "src"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--doctest-modules",
]
markers = [
    "unit: Unit tests (pure logic, no external dependencies)",
    "e2e: End-to-end tests with mocked SDKs",
    "live: Tests requiring real external services",
    "azure: Azure-specific tests",
]
